1. experimental support for Real==long double (#defines QUAD_PRECISION, scons quad=1). long double is not necessarily quadruple precision (not on IA32, at least) though, but is still more precise than double. (search "long double" on wikipedia) 2. Some preparation to compile with -std=c++0x (not yet functional) WRT <boost/shared_ptr.hpp> and boost::shared_ptr vs. <memory> and std::shared_ptr 3. Fixed in yade-multi. 4. New direct oofenm exporter in yade.eudoxos. 5. Don't fiddle with termios in non-interactive sessions.

+++ gui/qt3/SimulationController.cpp
@@ -411,7 +411,7 @@ void SimulationController::doUpdate(){
 	if(sbRefreshTime->value()!=refreshTime) sbRefreshTime->setValue(refreshTime);
 
 	char strStep[64];
-	snprintf(strStep,64,"step %g",Omega::instance().getTimeStep());
+	snprintf(strStep,64,"step %g",(double)Omega::instance().getTimeStep());
 	labelStep->setText(string(strStep));
 
 	changeSkipTimeStepper = false;
